1b/requirements.txt
PyMuPDF==1.23.8
sentence-transformers==2.7.0
torch>=2.1.1,<2.3
sumy==0.11.0
numpy<2.0
faiss-cpu==1.8.0
rank-bm25==0.2.2